Revert inclusion of previous row's line in line table
After further experimentation, it seems clear now that it is not correct to
include these lines from a previous row in the line table. While it is the right
way to map from instruction bytes to lines when between rows, it doesn't really
make sense from the debugging user's perspective to include such lines in
coverage. Most debugging users are stepping by lines, and if they try to stop at
the source line from the previous row, it will be _before_ the instruction range
we're analysing, and thus the variable won't be accessible there.

Reverts ecc0843d166111738636f40d4e7fca6c8802bf00